+++ picodi/_scopes.py
@@ -54,7 +54,7 @@ class Scope:
         """
         return None
 
-    def exit_decorator(self) -> None:
+    def exit_decorator(self, exc: BaseException | None = None) -> None:  # noqa: U100
         """
         Called before exiting a `inject` decorator.
         `close_local` will be called after this, e.g.:
@@ -69,8 +69,8 @@ class GlobalScope(Scope):
     Inherit this class for your custom global scope.
     """
 
-    def close_global(self) -> Awaitable:
-        return self.exit_stack.close()
+    def close_global(self, exc: BaseException | None = None) -> Awaitable:
+        return self.exit_stack.close(exc)
 
 
 class LocalScope(Scope):
@@ -113,9 +113,9 @@ class SingletonScope(GlobalScope):
     def set(self, key: Hashable, value: Any) -> None:
         self._store[key] = value
 
-    def close_global(self) -> Awaitable:
+    def close_global(self, exc: BaseException | None = None) -> Awaitable:
         self._store.clear()
-        return super().close_global()
+        return super().close_global(exc)
 
 
 class ParentCallScope(LocalScope):
@@ -135,7 +135,7 @@ class ParentCallScope(LocalScope):
         with self._lock:
             self._stack.get().append({})
 
-    def exit_decorator(self) -> None:
+    def exit_decorator(self, exc: BaseException | None = None) -> None:  # noqa: U100
         with self._lock:
             self._stack.get().pop()